- Decorators and font effects are now parsed by separate property parsers. - Decorator textures are now fetched from the decorator interface, instead of using the property source. - Decorators are now instanced when used in each element.
Due to the introduction of media queries, active sprites in a style sheets can now be changed at any time. Thus, we can no longer use the old trick of instancing all the decorators as soon as the style sheet is parsed. However, to ensure that performance is still good, I noticed that most of the work during decorator instancing was actually the parsing step. So now, we parse the decorator at the earliest opportunity, and then store the parsed properties of the decorator (DecoratorDeclaration). Then we instance the decorator later on once needed. The instancing step is actually quite fast once parsed, so it's okay now to delay (and possibly repeat) this step when needed.